+++ source/index.js
@@ -1,3 +1,0 @@
-import AnimateHeight from './AnimateHeight';
-
-export default AnimateHeight;